+++ src/ThemeUtils.php
@@ -53,7 +53,13 @@ class ThemeUtils
         self::$scssFolder = Path::join(self::getWebDir(), 'bundles', self::$themeFolder, self::$scssFolder);
 
         $scssStr = '';
-        $hash = hash('ripemd160', implode(' ,', $GLOBALS['ZERO_ONE_STYLES']));
+        $styles = implode(' ,', array_unique($GLOBALS['ZERO_ONE_STYLES']));
+
+        if (isset($GLOBALS['CUSTOM_STYLES'])) {
+            $styles .= ','.implode(' ,', array_unique($GLOBALS['CUSTOM_STYLES']));
+        }
+
+        $hash = hash('ripemd160', $styles);
         $objFile = new File('var/cache/zeroOne/scss/'.$hash.'.scss');
 
         if (!$objFile->exists()) {